Merge branch 'master' of https://github.com/SupSuper/OpenXcom into oXc_kL

Conflicts:
	bin/data/Language/en-GB.yml
	bin/data/Language/en-US.yml
	src/Battlescape/Camera.cpp
	src/Battlescape/Camera.h
	src/Battlescape/ProjectileFlyBState.cpp
	src/Battlescape/UnitFallBState.cpp
	src/Battlescape/UnitWalkBState.cpp
	src/Engine/Font.cpp
	src/Geoscape/GeoscapeState.cpp
	src/Geoscape/Globe.cpp
	src/Geoscape/Globe.h
	src/Resource/XcomResourcePack.cpp
	src/Savegame/Craft.cpp
	src/Savegame/Craft.h

+++ src/Savegame/SavedGame.cpp
@@ -881,7 +881,7 @@ void SavedGame::setGlobeLatitude(double lat)
  * Returns the current zoom level of the Geoscape globe.
  * @return, Zoom level.
  */
-int SavedGame::getGlobeZoom() const
+size_t SavedGame::getGlobeZoom() const
 {
 	return _globeZoom;
 }
@@ -890,7 +890,7 @@ int SavedGame::getGlobeZoom() const
  * Changes the current zoom level of the Geoscape globe.
  * @param zoom, Zoom level.
  */
-void SavedGame::setGlobeZoom(int zoom)
+void SavedGame::setGlobeZoom(size_t zoom)
 {
 	_globeZoom = zoom;
 }
